external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/shader-with-frag-depth.frag-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:6: 'gl_FragDepth' : undeclared identifier 

fragment shader that uses gl_FragDepth should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: fragment shader that uses gl_FragDepth should fail
default vertex shader
test fragment shaderPASS fragment shader that uses gl_FragDepth should fail
PASS successfullyParsed is true

TEST COMPLETE
